requirements.txt
# LLM
anthropic>=0.40.0

# Embeddings
sentence-transformers>=3.0.0

# Vector store
chromadb>=0.5.0

# UI
streamlit>=1.36.0

# Environment
python-dotenv>=1.0.0

# Data
pandas>=2.0.0
numpy>=1.26.0

# Testing
pytest>=7.0.0

# Web fetching (for knowledge base expansion)
requests>=2.32.0
beautifulsoup4>=4.12.0
